+++ Sources/MapboxDirections/Route.swift
@@ -37,6 +37,7 @@ extension Route: Equatable {
             lhs.distance == rhs.distance &&
             lhs.expectedTravelTime == rhs.expectedTravelTime &&
             lhs.speechLocale == rhs.speechLocale &&
+            lhs.responseContainsSpeechLocale == rhs.responseContainsSpeechLocale &&
             lhs.legs == rhs.legs &&
             lhs.shape == rhs.shape
     }